Restructed lrt directories to enable an ulnx build
Added requirement of '--enable-baremetal' flag on configure line for baremetal
builds.

Hello World baremetal tested. Nothing ulnx yet.
