+++ configure.ac
@@ -3,7 +3,7 @@ AC_INIT([EbbRT],[0.3],[sesa-developers@googlegroups.com])
 AC_CONFIG_AUX_DIR([build-aux])
 AM_INIT_AUTOMAKE([foreign subdir-objects -Wall -Werror -Wno-portability])
 #some random file which we assert exists in source tree
-AC_CONFIG_SRCDIR([src/lrt/boot.cpp])
+AC_CONFIG_SRCDIR([src/lrt/bare/boot.cpp])
 AC_CONFIG_HEADER([config.h])
 AC_CONFIG_FILES([Makefile])
 
@@ -15,6 +15,18 @@ AS_CASE([$host_cpu], [x86_64],
 [AC_MSG_ERROR([not on a supported architecture])])
 AM_CONDITIONAL([ARCH_X86_64], [test x$host_cpu == xx86_64])
 AM_CONDITIONAL([ARCH_POWERPC], [test x$host_cpu == xpowerpc])
+
+# Baremetal or Linux 
+AC_ARG_ENABLE([baremetal],
+  AS_HELP_STRING([--enable-baremetal], [Enable baremetal build]),
+  [lrt=bare],
+  [lrt=ulnx])
+
+AS_CASE([$lrt], [bare], [AC_DEFINE([LRT_BARE], [1], [Baremetal LRT])],
+		[ulnx], [AC_DEFINE([LRT_ULNX], [1], [Baremetal ULNX])])
+
+AM_CONDITIONAL(LRT_BARE, test x$lrt = xbare)
+AM_CONDITIONAL(LRT_ULNX, test x$lrt = xulnx)
 
 # Checks for programs.
 AM_PROG_AS
