+++ src/lrt/Makefile.am
@@ -1,12 +1,17 @@
-include src/lrt/arch/Makefile.am
+include src/lrt/bare/Makefile.am
+include src/lrt/ulnx/Makefile.am
 
 lrt_sources = \
-	$(lrt_arch_sources) \
-	src/lrt/boot.cpp \
 	src/lrt/boot.hpp \
 	src/lrt/console.hpp \
-	src/lrt/event.cpp \
 	src/lrt/event.hpp \
-	src/lrt/mem.cpp \
-	src/lrt/trans.cpp \
-	src/lrt/uart.cpp
+	src/lrt/mem.hpp \
+	src/lrt/trans.hpp 
+
+if LRT_ULNX
+lrt_sources += $(lrt_ulnx_sources)
+else
+if LRT_BARE
+lrt_sources += $(lrt_bare_sources)
+endif
+endif
