--- src/lrt/arch/Makefile.am
@@ -1,9 +1,0 @@
-include src/lrt/arch/x86_64/Makefile.am
-
-lrt_arch_sources = \
-	src/lrt/arch/boot.hpp \
-	src/lrt/arch/event.hpp
-
-if ARCH_X86_64
-lrt_arch_sources += $(lrt_arch_x86_64_sources)
-endif
