+++ src/lrt/bare/Makefile.am
@@ -1,0 +1,12 @@
+include src/lrt/bare/arch/Makefile.am
+
+lrt_bare_sources = \
+	$(lrt_arch_sources) \
+	src/lrt/bare/boot.cpp \
+	src/lrt/bare/boot.hpp \
+	src/lrt/bare/console.hpp \
+	src/lrt/bare/event.cpp \
+	src/lrt/bare/event.hpp \
+	src/lrt/bare/mem.cpp \
+	src/lrt/bare/trans.cpp \
+	src/lrt/bare/uart.cpp
