+++ src/lrt/bare/arch/Makefile.am
@@ -1,0 +1,13 @@
+include src/lrt/bare/arch/x86_64/Makefile.am
+
+lrt_arch_sources = \
+	src/lrt/bare/arch/boot.hpp \
+	src/lrt/bare/arch/event.hpp
+
+if ARCH_X86_64
+lrt_arch_sources += $(lrt_arch_x86_64_sources)
+endif
+
+
+
+
