+++ src/lrt/bare/arch/boot.hpp
@@ -3,7 +3,7 @@
 #endif
 
 #ifdef ARCH_X86_64
-#include "lrt/arch/x86_64/boot.hpp"
+#include "lrt/bare/arch/x86_64/boot.hpp"
 #else
 #error "Unsupported architecture"
 #endif
