+++ src/lrt/bare/arch/event.hpp
@@ -14,7 +14,7 @@ namespace ebbrt {
 }
 
 #ifdef ARCH_X86_64
-#include "lrt/arch/x86_64/event.hpp"
+#include "lrt/bare/arch/x86_64/event.hpp"
 #else
 #error "Unsupported Architecture"
 #endif
