+++ src/lrt/bare/arch/trans.hpp
@@ -11,7 +11,7 @@ namespace ebbrt {
 }
 
 #ifdef ARCH_X86_64
-#include "lrt/arch/x86_64/trans.hpp"
+#include "lrt/bare/arch/x86_64/trans.hpp"
 #else
 #error "Unsupported Architecture"
 #endif
