+++ src/lrt/bare/arch/trans_impl.hpp
@@ -1,9 +1,9 @@
-#ifndef EBBRT_LRT_TRANS_IMPL_HPP
+#ifndef EBBRT_LRT_BARE_TRANS_IMPL_HPP
 #error "Don't include this file directly"
 #endif
 
 #ifdef ARCH_X86_64
-#include "lrt/arch/x86_64/trans_impl.hpp"
+#include "lrt/bare/arch/x86_64/trans_impl.hpp"
 #else
 #error "Unsupported Architecture"
 #endif
