+++ src/lrt/bare/arch/x86_64/default_func.S
