+++ src/lrt/bare/arch/x86_64/smp.hpp
