+++ src/lrt/bare/arch/x86_64/trans.cpp
