+++ src/lrt/bare/boot.cpp
