+++ src/lrt/bare/console.hpp
@@ -1,0 +1,10 @@
+namespace ebbrt {
+  namespace lrt {
+    namespace console {
+      void init();
+      void write(char c);
+      int write(const char *str, int len);
+      void write(const char *str);
+    }
+  }
+}
