+++ src/lrt/bare/event_impl.hpp
@@ -1,0 +1,14 @@
+#ifndef EBBRT_LRT_BARE_EVENT_IMPL_HPP
+#define EBBRT_LRT_BARE_EVENT_IMPL_HPP
+
+#include "lrt/event.hpp"
+
+namespace ebbrt {
+  namespace lrt {
+    namespace event {
+      extern uintptr_t** altstack;
+    }
+  }
+}
+
+#endif
