+++ src/lrt/bare/mem_impl.hpp
@@ -1,0 +1,22 @@
+#ifndef EBBRT_LRT_BARE_MEM_IMPL_HPP
+#define EBBRT_LRT_BARE_MEM_IMPL_HPP
+
+#include "lrt/mem.hpp"
+
+namespace ebbrt {
+  namespace lrt {
+    namespace mem {
+      /**
+       * @brief Memory region for per-core allocation 
+       */
+      class Region {
+      public:
+        char* start;
+        char* current;
+        char* end;
+      };
+      extern Region* regions;
+    }
+  }
+}
+#endif
