+++ src/lrt/console.hpp
@@ -1,10 +1,5 @@
-namespace ebbrt {
-  namespace lrt {
-    namespace console {
-      void init();
-      void write(char c);
-      int write(const char *str, int len);
-      void write(const char *str);
-    }
-  }
-}
+#ifdef LRT_ULNX
+#include <src/lrt/ulnx/console.hpp>
+#elif LRT_BARE
+#include <src/lrt/bare/console.hpp>
+#endif
