+++ src/lrt/event_impl.hpp
@@ -1,14 +1,10 @@
 #ifndef EBBRT_LRT_EVENT_IMPL_HPP
 #define EBBRT_LRT_EVENT_IMPL_HPP
 
-#include "lrt/event.hpp"
-
-namespace ebbrt {
-  namespace lrt {
-    namespace event {
-      extern uintptr_t** altstack;
-    }
-  }
-}
+#ifdef LRT_ULNX
+#include <src/lrt/ulnx/event_impl.hpp>
+#elif LRT_BARE
+#include <src/lrt/bare/event_impl.hpp>
+#endif
 
 #endif
