+++ src/lrt/mem.hpp
@@ -1,41 +1,10 @@
 #ifndef EBBRT_LRT_MEM_HPP
 #define EBBRT_LRT_MEM_HPP
 
-#include "lrt/event.hpp"
-
-namespace ebbrt {
-  namespace lrt {
-    namespace mem {
-      extern char* mem_start;
-      /**
-       * @brief Initial memory configuration 
-       *
-       * @param num_cores
-       *
-       * @return 
-       */
-      bool init(unsigned num_cores);
-      /**
-       * @brief Lrt memory allocator
-       *
-       * @param size 
-       * @param loc
-       *
-       * @return 
-       */
-      void* malloc(size_t size, event::Location loc);
-      /**
-       * @brief Lrt memory alignment
-       *
-       * @param boundary
-       * @param size
-       * @param loc
-       *
-       * @return 
-       */
-      void* memalign(size_t boundary, size_t size, event::Location loc);
-    }
-  }
-}
+#ifdef LRT_ULNX
+#include <src/lrt/ulnx/mem.hpp>
+#elif LRT_BARE
+#include <src/lrt/bare/mem.hpp>
+#endif
 
 #endif
